+++ robit/core/counter.py
@@ -11,13 +11,13 @@ class Counter:
     def __repr__(self):
         return self.total
 
-    def increase(self, count: int = 1) -> None:
-        self.total += count
-
-    def decrease(self, count: int = 1) -> None:
-        self.total -= count
-
     def as_dict(self) -> dict:
         return {
             'count': self.total
         }
+
+    def decrease(self, count: int = 1) -> None:
+        self.total -= count
+
+    def increase(self, count: int = 1) -> None:
+        self.total += count