+++ robit/job/enums.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class JobExecutionTypeEnum(Enum):
+    THREAD = 'thread'
+    PROCESS = 'process'