+++ robit/job/group.py
@@ -1,4 +1,3 @@
-import threading
 from typing import Callable
 
 from robit.core.alert import Alert
@@ -7,24 +6,22 @@ from robit.core.health import Health
 from robit.core.id import Id
 from robit.job import Job
 from robit.core.name import Name
-from robit.core.status import Status
 
 
 class Group:
     def __init__(
             self,
-            name: str = 'default',
+            name: str = 'Unnamed Group',
             alert_method: Callable = None,
             alert_method_kwargs: dict = None,
             alert_health_threshold: float = 95.0,
     ):
-        self.id = Id()
-        self.name = Name(name)
-        self.health = Health()
-        self.status = Status()
-        self.clock = Clock()
+        self.id: Id = Id()
+        self.name: Name = Name(name)
+        self.health: Health = Health()
+        self.clock: Clock = Clock()
 
-        self.job_list = list()
+        self.job_list: list[Job] = list()
 
         if alert_method is not None:
             self.alert = Alert(
@@ -35,10 +32,7 @@ class Group:
         else:
             self.alert = None
 
-        self.thread = threading.Thread(target=self.run_job_list)
-        self.thread.daemon = True
-
-    def add_job(self, name: str, method, **kwargs):
+    def add_job(self, name: str, method: Callable, **kwargs):
         self.job_list.append(Job(name=name, method=method, **kwargs))
 
     def calculate_health(self):
@@ -58,30 +52,10 @@ class Group:
 
         return job_dict_full
 
-    def run_job_list(self):
-        while True:
-            for job in self.job_list:
-                job.run()
-
-            self.calculate_health()
-
-            if self.alert:
-                self.alert.check_health_threshold(f'Group "{self.name}"', self.health)
-
-    def restart(self):
-        pass
-
-    def start(self):
-        self.thread.start()
-
-    def stop(self):
-        pass
-
     def as_dict(self):
         return {
             'id': str(self.id),
             'name': str(self.name),
             'health': str(self.health),
             'jobs': self.convert_jobs_to_dict_list(),
-            'status': str(self.status),
         }