+++ robit/job/job.py
@@ -1,5 +1,4 @@
 import logging
-from time import sleep
 from typing import Callable, Optional
 
 from robit.core.alert import Alert
@@ -11,23 +10,25 @@ from robit.core.health import Health
 from robit.core.id import Id
 from robit.core.log import Log
 from robit.core.name import Name
-from robit.core.status import Status
-from robit.core.timer import Timer
+from robit.status import Status
+from robit.timer import Timer, timing_decorator
 
 
 class Job:
     def __init__(
             self,
             name: str,
-            method,
+            method: Callable,
             method_kwargs: Optional[dict] = None,
             cron: str = '* * * * *',
+            execution_type: str = 'thread',
             alert_method: Callable = None,
             alert_method_kwargs: dict = None,
     ):
         self.id = Id()
         self.name = Name(name)
         self.method = method
+        self.execution_type = execution_type
 
         if method_kwargs is None:
             self.method_kwargs = dict()
@@ -46,9 +47,7 @@ class Job:
             self.alert = None
 
         self.clock = Clock()
-
         self.timer = Timer()
-
         self.status = Status()
 
         self.success_count = Counter()
@@ -58,6 +57,14 @@ class Job:
         self.health = Health()
 
         self.result_log = Log(max_messages=200)
+
+    @timing_decorator
+    def execute_method(self):
+        if self.method_kwargs:
+            method_result = self.method(**self.method_kwargs)
+        else:
+            method_result = self.method()
+        return method_result
 
     @property
     def method_verbose(self):
@@ -76,40 +83,31 @@ class Job:
         return tz_now() > self.next_run_datetime
 
     def run(self):
-        if self.should_run():
-            self.set_next_run_datetime()
+        self.status.running()
 
-            logging.warning(f'STARTING: Job "{self.name}"')
+        logging.warning(f'STARTING: Job "{self.name}"')
 
-            self.status.set('run')
-            self.timer.start()
+        try:
+            method_result = self.execute_method()
+            logging.warning(f'SUCCESS: Job "{self.name}" completed')
 
-            try:
-                if self.method_kwargs:
-                    method_result = self.method(**self.method_kwargs)
-                else:
-                    method_result = self.method()
-                self.timer.stop()
-                logging.warning(f'SUCCESS: Job "{self.name}" completed')
-                self.success_count.increase()
-                self.health.add_positive()
-                if method_result:
-                    self.result_log.add_message(str(method_result))
-            except Exception as e:
-                self.status.set('error')
-                failed_message = f'ERROR: Job "{self.name}" failed on exception "{e}"'
-                logging.warning(failed_message)
-                self.failed_log.add_message(failed_message)
-                self.failed_count.increase()
-                self.health.add_negative()
+            self.success_count.increase()
+            self.health.add_positive()
 
-            if self.alert:
-                self.alert.check_health_threshold(f'Job "{self.name}"', self.health)
-        else:
-            if self.status.value != 'error':
-                self.status.set('queued')
+            if method_result:
+                self.result_log.add_message(str(method_result))
 
-            sleep(1)
+            self.status.queued()
+        except Exception as e:
+            self.status.error()
+            failed_message = f'ERROR: Job "{self.name}" failed on exception "{e}"'
+            logging.warning(failed_message)
+            self.failed_log.add_message(failed_message)
+            self.failed_count.increase()
+            self.health.add_negative()
+
+        if self.alert:
+            self.alert.check_health_threshold(f'Job "{self.name}"', self.health)
 
     def as_dict(self):
         return {