+++ robit/status/__init__.py
@@ -1,0 +1,1 @@
+from robit.status.status import Status