+++ robit/timer/__init__.py
@@ -1,0 +1,2 @@
+from robit.timer.timer import Timer
+from robit.timer.decorators import timing_decorator