+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## v0.3.1.a1
+- Separate processes for web server
+- Socket communication for web server and worker
+- Queue for ready jobs
+- Thread pool executor for jobs
+- Class structure for modules 
+- Cron Class Structure
+
 ## v0.3.0
 - Cron next date time bug fix